Fixed repeating word (#138)
* Fix misleading description about messaging service

As of current, the description of the messaging service setting on the network installation page suggests that the setting is dependant on the pure existence of a bungeecord or velocity proxy regardless of the usage of LPB/LPV.  I think it would be helpful to clear up that this indeed means a proxy which also runs LuckPerms, not just the existence of one.

* Minor typo on verbose page

While copying the first two sentences from the wiki I noticed a minor typo in the very first word. I am a perfectionist, don't judge me.

* Update Network-Installation.md

* Removed double

Two placeholders descriptions used 'is currently' twice in the same sentence. I removed one double.

* Missing Letter

Thanks to ThomasWega#3382 on Discord for reporting this, added to my stuff since I already have a corrective PR open so why not.

Co-authored-by: Emily <35617540+emilyy-dev@users.noreply.github.com>